Bugfixes and QoL
Updated summary and benchmark panels to hide group average and deviation rows when the average value is zero or null. Added support to change names without having to change roomsize everytime, no longer able for rooms to be saved as null, as you have to enter a name now. No longer able to double save rooms.